Fix Profile, SongGen, and SongSearch
Fix the error introduced in the last amy2 pull that removed the scroll bar for overflowing song lyrics.
Fixed error in SongGeneration that made the makesong result display with no line breaks
Fix Profile so that history title is constant and doesn't scroll with the entries. Added title for profile info